Lowered nonce multiplier
Nonce associated to an API key must be increasing in time. A high multiplier can be source of bug that is hard to identify as API response will not explain the issue. By dimishing the nonce multiplier, we reduce risk that an API call returns a error response when we switch to another nonce scheme (usually higher than the current multiplier).