Fix issue when only BTC/ETH address and no price
Issue: when BTC or ETH are not held on exchanges but held on ETH/BTC addresses only, the exchanges get_balance() would not provide with prices for BTC or ETH since they would only request the prices for holdings at the exchange.
Fix: Using ref_api, we call the exchange's API to get the price of the BTC/ETH address even if at the exchange, no position in ETH or BTC is held.