Fix issue in balance when fiat positions exist
Previously, when a fiat position existed, the amount of fiat would count towards the total amount of crypto. For example, if you had 100 USD in your portfolio and used BTC as base_fiat, your total BTC equivalent would be accrued by 100 BTC. Now, the fiat position is taken into account in the total portfolio in fiat (as long as it is in the same currency as base_fiat) but ignored in the total in base_crypto.